Remove unnecessary gasprice updater logic and tests
