+++ plugin/evm/vm.go
@@ -218,17 +218,6 @@ type VM struct {
 	warpBackend warp.WarpBackend
 }
 
-/*
- ******************************************************************************
- ********************************* Snowman API ********************************
- ******************************************************************************
- */
-
-// implements SnowmanPlusPlusVM interface
-func (vm *VM) GetActivationTime() time.Time {
-	return time.Unix(vm.chainConfig.SubnetEVMTimestamp.Int64(), 0)
-}
-
 // Initialize implements the snowman.ChainVM interface
 func (vm *VM) Initialize(
 	_ context.Context,
@@ -390,6 +379,7 @@ func (vm *VM) Initialize(
 	vm.chainConfig = g.Config
 	vm.networkID = vm.ethConfig.NetworkId
 
+	// TODO: remove SkipSubnetEVMUpgradeCheck after next network upgrade
 	if !vm.config.SkipSubnetEVMUpgradeCheck {
 		// check that subnetEVM upgrade is enabled from genesis before upgradeBytes
 		if !vm.chainConfig.IsSubnetEVM(common.Big0) {
@@ -481,9 +471,6 @@ func (vm *VM) initializeChain(lastAcceptedHash common.Hash, ethConfig ethconfig.
 	vm.txPool = vm.eth.TxPool()
 	vm.blockChain = vm.eth.BlockChain()
 	vm.miner = vm.eth.Miner()
-
-	// start goroutines to update the tx pool gas minimum gas price when upgrades go into effect
-	vm.handleGasPriceUpdates()
 
 	vm.eth.Start()
 	return vm.initChainState(vm.blockChain.LastAcceptedBlock())
